Update model: group and variant fix tambahan jika kondisi awal = null

+++ app/Models/Variant.php
@@ -121,7 +121,9 @@ class Variant extends Model
             Storage::disk($disk)->put($destination_path . '/' . $filename, $image->stream());
 
             // 3. Delete the previous image, if there was one.
-            Storage::disk($disk)->delete(Str::replaceFirst('uploads/', '', $this->{$attribute_name}));
+            if ($this->{$attribute_name}) {
+                Storage::disk($disk)->delete(Str::replaceFirst('uploads/', '', $this->{$attribute_name}));
+            }
 
             // 4. Save the public path to the database
             // but first, remove "public/" from the path, since we're pointing to it
